# service/openapi/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Atlas Enterprise Agent API
  version: 0.1.0
  description: 单企业风险排查与 DOCX 报告生成 V1 API
servers:
  - url: http://localhost:8080
paths:
...
  /api/tasks:
    get:
      summary: List operator tasks with workflow summaries
      operationId: listTasks
      parameters:
        - name: query
          in: query
          required: false
          description: Case-insensitive match on task number, company query, or prompt
          schema:
            type: string
            maxLength: 256
        - name: status
          in: query
          required: false
          description: Repeat this parameter to include multiple task statuses
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/TaskStatus'
        - name: operator_id
          in: query
          required: false
          schema:
            type: string
            maxLength: 64
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          required: false
          description: Opaque cursor returned by the previous page
          schema:
            type: string
      responses:
        '200':
          description: Tasks ordered by updated time and task ID descending
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskListPage'
        '400':
          $ref: '#/components/responses/InvalidRequest'
    post:
      summary: 创建风险报告更新任务
      operationId: createTask
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/OperatorId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '202':
          description: 任务已接受
          headers:
            Location:
              schema:
                type: string
            X-Trace-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/InvalidRequest'
  /api/tasks/{taskId}:
    get:
      summary: 查询任务
      operationId: getTask
      parameters:
        - $ref: '#/components/parameters/TaskId'
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: 任务不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
...
    TaskId:
      name: taskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
    OperatorId:
      name: X-Operator-Id
      in: header
      required: false
      schema:
        type: string
        default: local-operator
...
  responses:
    InvalidRequest:
      description: 请求不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    CreateTaskRequest:
      type: object
      additionalProperties: false
      required:
        - prompt
        - company_query
      properties:
        prompt:
          type: string
          maxLength: 4000
        company_query:
          type: string
          maxLength: 256
    Task:
      type: object
      required:
        - task_id
        - task_no
        - intent
        - status
        - original_prompt
        - company_query
        - operator_id
        - created_at
        - updated_at
        - events_url
      properties:
        task_id:
          type: string
          format: uuid
        task_no:
          type: string
        intent:
          type: string
          const: RISK_REPORT_UPDATE
        status:
          $ref: '#/components/schemas/TaskStatus'
        original_prompt:
          type: string
        company_query:
          type: string
        previous_report_file_id:
          type:
            - string
            - 'null'
          description: 仅历史任务兼容，新任务不再使用
        atlas_company_id:
          type:
            - string
            - 'null'
          format: uuid
        current_step:
          type:
            - string
            - 'null'
        failed_step:
          type:
            - string
            - 'null'
        error_code:
          type:
            - string
            - 'null'
        operator_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type:
            - string
            - 'null'
          format: date-time
        events_url:
          type: string
    TaskStatus:
      type: string
      enum:
        - CREATED
        - RESOLVING_SUBJECT
        - WAITING_SUBJECT_CONFIRMATION
        - LOADING_PREVIOUS_REPORT
        - COLLECTING_STRUCTURED_DATA
        - SEARCHING_PUBLIC_INTELLIGENCE
        - CALCULATING_RISK
        - WAITING_OPERATOR_CONFIRMATION
        - GENERATING_REPORT
        - COMPLETED
        - SOURCE_FAILED
        - MODEL_FAILED
        - REPORT_FAILED
        - CANCELLED
    TaskListPage:
      type: object
      required:
        - items
        - has_more
        - page_size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TaskListItem'
        next_cursor:
          type:
            - string
            - 'null'
        has_more:
          type: boolean
        page_size:
          type: integer
          minimum: 1
          maximum: 100
    TaskListItem:
      type: object
      required:
        - task
        - company_name
        - evidence_progress
        - confirmation_state
        - readiness_blockers
        - next_action
      properties:
        task:
          $ref: '#/components/schemas/Task'
        company_name:
          type: string
        risk:
          anyOf:
            - $ref: '#/components/schemas/TaskRiskSummary'
            - type: 'null'
        evidence_progress:
          $ref: '#/components/schemas/EvidenceProgress'
        confirmation_state:
          type: string
          enum: [NOT_READY, PENDING, VALID, STALE]
        readiness_blockers:
          type: array
          items:
            type: string
            enum:
              - DATA_SNAPSHOT_MISSING
              - RISK_SCORE_MISSING
              - RISK_SCORE_NOT_FROM_LATEST_DATA
              - UNVERIFIED_EVIDENCE
        next_action:
          type: string
          enum:
            - EXECUTE_TASK
            - CONFIRM_SUBJECT
            - REVIEW_EVIDENCE
            - CALCULATE_RISK
            - CONFIRM_REVIEW
            - GENERATE_REPORT
            - RETRY_TASK
            - RETRY_REPORT
            - WAIT
            - DOWNLOAD_REPORT
            - NONE
        latest_confirmation:
          anyOf:
            - $ref: '#/components/schemas/TaskConfirmationSummary'
            - type: 'null'
        latest_report:
          anyOf:
            - $ref: '#/components/schemas/TaskReportSummary'
            - type: 'null'
    TaskRiskSummary:
      type: object
      required:
        - score_snapshot_id
        - original_score
        - manual_score
        - original_risk_level
        - manual_risk_level
        - calculated_at
      properties:
        score_snapshot_id:
          type: string
          format: uuid
        original_score:
          type: number
          minimum: 0
          maximum: 10
        manual_score:
          type: number
          minimum: 0
          maximum: 10
        original_risk_level:
          type: string
        manual_risk_level:
          type: string
        calculated_at:
          type: string
          format: date-time
    TaskConfirmationSummary:
      type: object
      required:
        - confirmation_id
        - operator_id
        - confirmed_evidence_count
        - rejected_evidence_count
        - confirmed_at
      properties:
        confirmation_id:
          type: string
          format: uuid
        operator_id:
          type: string
        confirmed_evidence_count:
          type: integer
          minimum: 0
        rejected_evidence_count:
          type: integer
          minimum: 0
        confirmed_at:
          type: string
          format: date-time
    TaskReportSummary:
      type: object
      required:
        - report_id
        - report_version_no
        - status
        - template_version
      properties:
        report_id:
          type: string
          format: uuid
        report_version_no:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [GENERATING, GENERATED, FAILED, SUPERSEDED]
        template_version:
          type: string
        generated_at:
          type:
            - string
            - 'null'
          format: date-time
...
    EvidenceProgress:
      type: object
      required:
        - total
        - confirmed
        - rejected
        - unverified
        - captured_content
        - failed_content_capture
        - by_risk_type
      properties:
        total:
          type: integer
          minimum: 0
        confirmed:
          type: integer
          minimum: 0
        rejected:
          type: integer
          minimum: 0
        unverified:
          type: integer
          minimum: 0
        captured_content:
          type: integer
          minimum: 0
        failed_content_capture:
          type: integer
          minimum: 0
        by_risk_type:
          type: object
          additionalProperties:
            type: integer
            minimum: 0
...
    Error:
      type: object
      required:
        - code
        - message
        - trace_id
        - retryable
        - details
        - occurred_at
      properties:
        code:
          type: string
        message:
          type: string
        task_id:
          type:
            - string
            - 'null'
          format: uuid
        trace_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
        occurred_at:
          type: string
          format: date-time